==> analysis/5_calculating_zscores.Rmd <==
---
title: "5_calculating_zscores"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# load packages
```{r message=FALSE, warning=FALSE}
library(tidyverse)
library(here)
library(skimr)
library(naniar)
```

# read in data

```{r message=FALSE, warning=FALSE}
data <- read_csv(here::here("data", "adult_child_combined", "clean_no_bl_outliers.csv"))
```

# first test on one participant 

```{r}
pp401 <- data %>%
  filter(pp_no == "pp401")

#make it wide

widepp401 <- pp401 %>%
  pivot_wider(names_from = muscle, values_from = rms)

#create a new variable called zbrow and zcheek using the scale function and specify center = TRUE and scale = TRUE

widepp401$zbrow = scale(widepp401$brow, center = TRUE, scale = TRUE)

widepp401$zcheek = scale(widepp401$cheek, center = TRUE, scale = TRUE)

#check that the mean of the new zbrow and zcheek columns is 0 

summary(widepp401)

```


# now do all the data 

- make data wide, use mutate to make new z variables 

```{r}
data_zscore <- data %>%
  pivot_wider(names_from = muscle, values_from = rms) %>%
  group_by(pp_no) %>%
  mutate(zbrow = scale(brow, center = TRUE, scale = TRUE)) %>%
  mutate(zcheek = scale(cheek, center = TRUE, scale = TRUE))

```



summary(data_zscore)


# calculating difference scores 

- calculate differences scores from baseline
- need to make BIN wide to allow for calculations across columns
- do brow and cheek separately 

# BROW DATA

## add a muscle column
```{r}
brow_z <- data_zscore %>%
  mutate(muscle = "brow") %>%
  select(pp_no, model, emotion, bin, trial, muscle, zbrow)
```

## make the bin column wide and rename bin_0 as BL

```{r}
brow_z_wide <- brow_z %>%
  pivot_wider(names_from = bin, values_from = zbrow) %>%
  rename(BL = bin_0)
```

## mutate diff scores 

- use wide columns to calculate the difference between each bin column and BL, creating a new set of columns starting with "diff", drop BL column and all columns starting with bin (i.e. raw z scores)

```{r}
brow_z_diff <- brow_z_wide %>%
  mutate(diff_bin1 = bin_1 - BL,
         diff_bin2 = bin_2 - BL,
         diff_bin3 = bin_3 - BL,
         diff_bin4 = bin_4 - BL,
         diff_bin5 = bin_5 - BL,
         diff_bin6 = bin_6 - BL,
         diff_bin7 = bin_7 - BL,
         diff_bin8 = bin_8 - BL,
         diff_bin9 = bin_9 - BL,
         diff_bin10 = bin_10 - BL) %>%
  select(-BL, - starts_with("bin"))
```

## make data long again 

```{r}
brow_z_diff_long <- brow_z_diff %>%
  pivot_longer(names_to = "bin", values_to = "zdiff", diff_bin1:diff_bin10)
```

## check on missing values 

```{r}
vis_miss(brow_z_diff_long)
```

#CHEEK 

## add a muscle column 

```{r}
cheek_z <- data_zscore %>%
  mutate(muscle = "cheek") %>%
  select(pp_no, model, emotion, bin, trial, muscle, zcheek)
```

# make the bin column wide and rename bin_0 as BL

```{r}
cheek_z_wide <- cheek_z %>%
  pivot_wider(names_from = bin, values_from = zcheek) %>%
  rename(BL = bin_0)
```

Use wide columns to calcuate the difference between each bin column and BL, creating a new set of columns starting with “diff”, drop BL column and all columns starting with bin (i.e. raw z scores)
```{r}
cheek_z_diff <- cheek_z_wide %>%
  mutate(diff_bin1 = bin_1 - BL,
         diff_bin2 = bin_2 - BL,
         diff_bin3 = bin_3 - BL,
         diff_bin4 = bin_4 - BL,
         diff_bin5 = bin_5 - BL,
         diff_bin6 = bin_6 - BL,
         diff_bin7 = bin_7 - BL,
         diff_bin8 = bin_8 - BL,
         diff_bin9 = bin_9 - BL,
         diff_bin10 = bin_10 - BL) %>%
  select(-BL, - starts_with("bin"))
```

## make data long again 

```{r}
cheek_z_diff_long <- cheek_z_diff %>%
  pivot_longer(names_to = "bin", values_to = "zdiff", diff_bin1:diff_bin10)
```

# bind cheek and brow together

```{r}
zdiff_cheek_brow <- bind_rows(cheek_z_diff_long, brow_z_diff_long) %>%
  arrange(pp_no, model, emotion, trial, muscle, bin)
```

# write to csv

```{r}
zdiff_cheek_brow %>%
  write_csv(here("data", "adult_child_combined", "zdiff_cheek_brow.csv"))
```
