==> analysis/11_plotting_cheek.Rmd <==
---
title: "11_plotting_cheek"
output: html_document
---

# load packages
```{r message=FALSE, warning=FALSE}
library(tidyverse)
library(here)
library(lme4)
library(lmerTest)
library(broom.mixed)
library(ggplot2) 
library(ggeasy)
library(dplyr)
library(RColorBrewer)
library(papaja)
library(lubridate)
```

# read in data 
```{r}
df <- read_csv(here::here("data", "adult_child_combined", "zdiff_binscreened2.csv"))
```

# fix data types

```{r}
df$emotion <- as.factor(df$emotion)

levels(df$emotion)


df <- df %>% mutate_if(is.character, as.factor)

glimpse(df)
```


#CHILD HA CHEEK

```{r}
plot_child_cheek <- df %>%
   filter(emo %in% c("happy", "angry", "fear", "sad")) %>%
  filter(emotion %in% c("121", "323", "424", "525")) %>%
  filter(muscle == "cheek") %>%
  filter(model == "child") %>%
  arrange(pp_no, emotion, emo, trial, bin)

glimpse(plot_child_cheek)
```

#make sure it is only the relevant emotions
```{r}

plot_child_cheek$emo <- fct_relevel(plot_child_cheek$emo, c("happy", "angry", "fear", "sad"))

levels(plot_child_cheek$emo)


plot_child_cheek$bin <- fct_relevel(plot_child_cheek$bin, c("diff_bin1", "diff_bin2", "diff_bin3", "diff_bin4", "diff_bin5", "diff_bin6", "diff_bin7", "diff_bin8", "diff_bin9", "diff_bin10"))

levels(plot_child_cheek$bin)
```

#create a new bin column and rename bins
```{r}

plot_child_cheek$time_bin <- plot_child_cheek$bin


plot_child_cheek <- plot_child_cheek %>%
  mutate(time_bin = recode(bin, "diff_bin1" = "100",
                                "diff_bin2" = "200", 
                                "diff_bin3" = "300", 
                                "diff_bin4" = "400", 
                              "diff_bin5" = "500", 
                              "diff_bin6" = "600", 
                              "diff_bin7" = "700", 
                              "diff_bin8" = "800",
                              "diff_bin9" = "900",
                              "diff_bin10" = "1000", .default = "NA"))
                           
```



#plot cheek activity 

```{r}
plot_data <- plot_child_cheek %>%
  filter(model == "child") %>%
   filter(emo %in% c("happy", "angry", "fear", "sad")) %>%
  filter(emotion %in% c("121", "323", "424", "525")) %>%
  filter(muscle == "cheek") %>%
  group_by(emo, time_bin) %>%
  summarise(meanZ = mean(zdiff, na.rm = TRUE), sd_RMS = sd(zdiff, na.rm = TRUE), n = n(), stderr = sd_RMS/sqrt(n)) %>%
  ggplot(aes(x = time_bin, y = meanZ, group = emo, colour = emo)) +
  geom_point() +
  geom_line() +
  scale_colour_grey(start = 0.0, end = 0.8) +
  geom_errorbar(aes(ymin = meanZ-stderr, ymax = meanZ+stderr), size = .4, width = .4) +
  theme_classic() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), panel.background = element_rect(colour = "black", size = 1)) +
  scale_y_continuous(limits = c(-0.2, 0.2)) +
  labs(y = "Mean difference in cheek activity 
(baseline corrected z-score)", x = "Time post-stimulus presentation onset (ms)") +
 easy_x_axis_title_size(size = 14) +
 easy_y_axis_title_size(size = 14)

print(plot_data)

```

```{r}
ggsave("child_cheek_plot.png")
```


#bar graph 

#read in data 

```{r}
plot_child <- df %>%
  filter(emo %in% c("happy", "angry", "fear", "sad")) %>%
  filter(model == "child") %>%
  arrange(pp_no, emotion, emo, muscle, trial, bin)

glimpse(plot_child)
```



```{r}
plot_data <- plot_child %>%
  filter(model == "child") %>%
  filter(emo %in% c("happy", "angry", "fear", "sad")) %>%
  filter(muscle == c("cheek", "brow")) %>%
  group_by(emo, muscle) %>%
  summarise(meanZ = mean(zdiff, na.rm = TRUE), sd_RMS = sd(zdiff, na.rm = TRUE), n = n(), stderr = sd_RMS/sqrt(n)) %>%
  ggplot(aes(x = emo, y = meanZ, fill = emo)) +
  geom_col(position = "dodge") +
  theme_apa(base_size = 14) +
  scale_fill_grey(start = 0.3, end = 0.7) +
  geom_errorbar(aes(ymin = meanZ-stderr, ymax = meanZ+stderr), size = .4, width = .4, position = position_dodge(.9)) +
  theme_classic() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), panel.background = element_rect(colour =    "black", size = 1)) +
  scale_y_continuous(limits = c(-0.1, 0.1)) +
  facet_wrap(~muscle) +
  labs(y = "Mean z-score difference (RMS) from baseline
(averaged across bins)", x = "Emotion") +
  easy_remove_legend() +
  easy_x_axis_title_size(size = 14) +
  easy_y_axis_title_size(size = 14) +
  easy_x_axis_labels_size(size = 12) +
  easy_y_axis_labels_size(size = 12)

print(plot_data)
```

```{r}
ggsave("child_bar_plot.png")
```



#ADULT HA CHEEK 

```{r}
plot_adult_cheek <- df %>%
  filter(emo %in% c("happy", "angry", "fear", "sad")) %>%
  filter(muscle == "cheek") %>%
  filter(model == "adult") %>%
  arrange(pp_no, emotion, emo, trial, bin)

glimpse(plot_adult_cheek)
```

#make sure it is only the relevant emotions
```{r}

plot_adult_cheek$emo <- fct_relevel(plot_adult_cheek$emo, c("happy", "angry", "fear", "sad"))

levels(plot_adult_cheek$emo)


plot_adult_cheek$bin <- fct_relevel(plot_adult_cheek$bin, c("diff_bin1", "diff_bin2", "diff_bin3", "diff_bin4", "diff_bin5", "diff_bin6", "diff_bin7", "diff_bin8", "diff_bin9", "diff_bin10"))

levels(plot_adult_cheek$bin)
```


#create new column and rename bins
```{r}
plot_adult_cheek$time_bin <- plot_adult_cheek$bin


plot_adult_cheek <- plot_adult_cheek %>%
  mutate(time_bin = recode(bin, "diff_bin1" = "100",
                                "diff_bin2" = "200", 
                                "diff_bin3" = "300", 
                                "diff_bin4" = "400", 
                              "diff_bin5" = "500", 
                              "diff_bin6" = "600", 
                              "diff_bin7" = "700", 
                              "diff_bin8" = "800",
                              "diff_bin9" = "900",
                              "diff_bin10" = "1000", .default = "NA"))
```


#plot adult cheek 
```{r}
plot_data <- plot_adult_cheek %>%
  filter(model == "adult") %>%
  filter(emo %in% c("happy", "angry", "fear", "sad")) %>%
  filter(muscle == "cheek") %>%
  group_by(emo, time_bin) %>%
  summarise(meanZ = mean(zdiff, na.rm = TRUE), sd_RMS = sd(zdiff, na.rm = TRUE), n = n(), stderr = sd_RMS/sqrt(n)) %>%
  ggplot(aes(x = time_bin, y = meanZ, group = emo, colour = emo)) +
  geom_point() +
  geom_line() +
  geom_errorbar(aes(ymin = meanZ-stderr, ymax = meanZ+stderr), size = .4, width = .4) +
  theme_classic() +
  scale_colour_grey(start = 0.0, end = 0.8) +
  theme_apa() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), panel.background = element_rect(colour =    "black", size = 1)) +
  scale_y_continuous(limits = c(-0.15, 0.25)) + 
  labs(y = "Mean difference in cheek activity 
(baseline corrected z-score)", x = "Time post-stimulus presentation onset (ms)") +
easy_x_axis_title_size(size = 14) +
easy_y_axis_title_size(size = 14)

print(plot_data)
```


```{r}
plot_data <- plot_adult_cheek %>%
  filter(model == "adult") %>%
  filter(emo %in% c("happy", "angry")) %>%
  filter(muscle == "cheek") %>%
  group_by(emo, time_bin) %>%
  summarise(meanZ = mean(zdiff, na.rm = TRUE), sd_RMS = sd(zdiff, na.rm = TRUE), n = n(), stderr = sd_RMS/sqrt(n)) %>%
  ggplot(aes(x = time_bin, y = meanZ, group = emo, colour = emo)) +
  geom_point() +
  geom_line() +
  geom_errorbar(aes(ymin = meanZ-stderr, ymax = meanZ+stderr), size = .4, width = .4) +
  theme_classic() +
  scale_colour_grey(start = 0.0, end = 0.6) +
  theme_apa() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), panel.background = element_rect(colour =    "black", size = 1)) +
  scale_y_continuous(limits = c(-0.15, 0.4)) + 
  labs(y = "Mean difference in cheek activity 
(baseline corrected z-score)", x = "Time post-stimulus presentation onset (ms)") +
easy_x_axis_title_size(size = 14) +
easy_y_axis_title_size(size = 14)

print(plot_data)

ggsave("happyangryadult.png")

```









```{r}
plot_adult <- df %>%
  filter(emo %in% c("happy", "angry", "fear", "sad")) %>%
  filter(model == "adult") %>%
  arrange(pp_no, emotion, emo, muscle, trial, bin)

glimpse(plot_adult)
```




```{r}
plot_data <- plot_adult %>%
  filter(model == "adult") %>%
  filter(emo %in% c("happy", "angry", "fear", "sad")) %>%
  filter(muscle == c("cheek", "brow")) %>%
  group_by(emo, muscle) %>%
  summarise(meanZ = mean(zdiff, na.rm = TRUE), sd_RMS = sd(zdiff, na.rm = TRUE), n = n(), stderr = sd_RMS/sqrt(n)) %>%
  ggplot(aes(x = emo, y = meanZ, fill = emo)) +
  geom_col(position = "dodge") +
  theme_apa(base_size = 14) +
  scale_fill_grey(start = 0.3, end = 0.7) +
  geom_errorbar(aes(ymin = meanZ-stderr, ymax = meanZ+stderr), size = .5, width = .4, position = position_dodge(.9)) +
  theme_classic() +
  facet_wrap(~muscle) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), panel.background = element_rect(colour =    "black", size = 1)) +
  scale_y_continuous(limits = c(-0.1, 0.1)) +
  labs(y = "Mean z-score difference (RMS) from baseline
(averaged across bins)", x = "Emotion") +
    easy_remove_legend() +
  easy_x_axis_title_size(size = 14) +
  easy_y_axis_title_size(size = 14) +
  easy_x_axis_labels_size(size = 12) +
  easy_y_axis_labels_size(size = 12)

print(plot_data)
```
