==> analysis/13_adult_child_analysis.Rmd <==
---
title: "13_adult_child_analyses"
output: html_document
---

# load packages
```{r message=FALSE, warning=FALSE}
library(tidyverse)
library(here)
library(lme4)
library(lmerTest)
library(broom.mixed)
library(pixiedust)
library(beepr)
library(ggplot2) 
```

Goal here is to work out whether it is possible to compare responses to child and adult faces i.e. to include model in the model. 

# read in data 
```{r}
df <- read_csv(here::here("data", "adult_child_combined", "zdiff_binscreened2.csv"))
```

# fix data types

add log modulus

```{r}

df <- df %>% mutate_if(is.character, as.factor) %>%
  mutate(log_modulus = sign(zdiff) * log(1+abs(zdiff)))
  

glimpse(df)
```

# CHEEK happy-angry 

### Make cheek HA df 

just happy/angry face for cheek

```{r}
HA_cheek <- df %>%
  filter(emo %in% c("happy", "angry")) %>%
  filter(muscle == "cheek") %>%
  arrange(pp_no, model, trial, bin)

HA_cheek$emo <- fct_relevel(HA_cheek$emo, c("happy", "angry"))

HA_cheek$bin <- fct_relevel(HA_cheek$bin, c("diff_bin1", "diff_bin2", "diff_bin3", "diff_bin4", "diff_bin5", "diff_bin6", "diff_bin7", "diff_bin8", "diff_bin9", "diff_bin10"))

levels(HA_cheek$bin)
```

add "model" to best fit from 9_cheek_analysis 

## model 1 (slopes emo)

This model adds model to best fit model (including slopes for emo)

> adult_cheek_lm_2 <- lmer(log_modulus ~ emo + bin + emo*bin + (1 + emo|pp_no) , data = HA_adult_cheek, REML = FALSE)

```{r}
cheek_lm_1 <- lmer(log_modulus ~ emo + bin + model + 
                    emo*bin + bin*model +
                    emo*bin*model +
                    (1 + emo|pp_no) , 
                  data = HA_cheek, REML = FALSE)

beepr::beep()
```

Error- Doesn't converge... remove slope for emo, just include intercept. 

## model 2 (just intercepts participant)
```{r}
cheek_lm_2 <- lmer(log_modulus ~ emo + bin + model + 
                    emo*bin + bin*model +
                    emo*bin*model +
                    (1|pp_no) , 
                  data = HA_cheek, REML = FALSE)

beepr::beep()
```


# compare models with and without slopes

hmmm weird that model with slope (lm_1) has lower AIC than model without (lm_2) even though lm_1 doesn't converge.  Model with slope does better than model without. 

```{r}
AIC(cheek_lm_1)
AIC(cheek_lm_2)

anova(cheek_lm_1, cheek_lm_2)
```

### get summary/tidydf

```{r}
summary(cheek_lm_2)

tidy_cheek2 <- tidy(cheek_lm_2)
```


# BROW happy angry

### Make brow HA df 

just happy/angry face for brow

```{r}
HA_brow <- df %>%
  filter(emo %in% c("happy", "angry")) %>%
  filter(muscle == "brow") %>%
  arrange(pp_no, model, trial, bin)

HA_brow$emo <- fct_relevel(HA_brow$emo, c("happy", "angry"))

HA_brow$bin <- fct_relevel(HA_brow$bin, c("diff_bin1", "diff_bin2", "diff_bin3", "diff_bin4", "diff_bin5", "diff_bin6", "diff_bin7", "diff_bin8", "diff_bin9", "diff_bin10"))

levels(HA_brow$bin)
```

add "model" to best fit from 10_brow_analysis 

## model 1 (slopes emo)

This model adds model to best fit model (including slopes for emo)

> child_brow_lm_2 <- lmer(log_modulus_new ~ emo + bin + emo*bin + (1 + emo|pp_no), data = HA_child_brow, REML = FALSE)

```{r}
brow_lm_1 <- lmer(log_modulus ~ emo + bin + model + 
                    emo*bin + bin*model +
                    emo*bin*model +
                    (1 + emo|pp_no) , 
                  data = HA_brow, REML = FALSE)

beepr::beep()
```

This one DOES converge... does it do better than model without slopes for emo? 

## model 2 (just intercepts participant)
```{r}
brow_lm_2 <- lmer(log_modulus ~ emo + bin + model + 
                    emo*bin + bin*model +
                    emo*bin*model +
                    (1|pp_no) , 
                  data = HA_brow, REML = FALSE)

beepr::beep()
```

# compare models with and without slopes

yes model with slope does better than model without. 

```{r}
AIC(brow_lm_1)
AIC(brow_lm_2)

anova(brow_lm_1, brow_lm_2)
```


### get summary/tidydf
```{r}
summary(cheek_lm_2)

tidy_cheek2 <- tidy(cheek_lm_2)
```
